+++ packages/plugin-rows/src/editor.tsx
@@ -11,18 +11,12 @@ export const RowsEditor = (
   props: StatefulPluginEditorProps<typeof rowsState>
 ) => {
   const rows = props.state
-
   return (
     <RowsContainer>
       {rows.items.map((row, index) => {
         return (
           <div key={row.id} style={{ position: 'relative' }}>
-            <Row
-              state={props.state}
-              editable={props.editable}
-              focused={props.focused}
-              index={index}
-            />
+            <Row {...props} index={index} />
           </div>
         )
       })}