+++ packages/plugin-video/src/editor.tsx
@@ -9,15 +9,14 @@ import * as React from 'react'
 import { VideoRenderer } from './renderer'
 import { videoState } from '.'
 
-export const VideoEditor = ({
-  editable,
-  focused,
-  state
-}: StatefulPluginEditorProps<typeof videoState>) => {
+export const VideoEditor = (
+  props: StatefulPluginEditorProps<typeof videoState>
+) => {
+  const { editable, focused, state } = props
   return (
     <React.Fragment>
       <ContainerWithConfigButton>
-        <VideoRenderer state={state} disableCursorEvents={editable} />
+        <VideoRenderer {...props} disableCursorEvents={editable} />
       </ContainerWithConfigButton>
       {focused ? (
         <Overlay>