+++ packages/ui-editor/src/components/textarea.tsx
@@ -1,17 +1,15 @@
 import * as React from 'react'
 
-import { createEditorUiElementTheme, EditorThemeProps, styled } from '../theme'
+import { createEditorUiTheme, EditorThemeProps, styled } from '../theme'
 
-export const createTextareaTheme = createEditorUiElementTheme<TextareaTheme>(
-  theme => {
-    return {
-      backgroundColor: 'transparent',
-      color: theme.color,
-      borderColor: theme.color,
-      highlightColor: theme.primary.background
-    }
+export const createTextareaTheme = createEditorUiTheme<TextareaTheme>(theme => {
+  return {
+    backgroundColor: 'transparent',
+    color: theme.color,
+    borderColor: theme.color,
+    highlightColor: theme.primary.background
   }
-)
+})
 const TextareaLabel = styled.label((props: EditorThemeProps) => {
   const theme = createTextareaTheme('textarea', props.theme)