+++ packages/ui-renderer/src/theme.ts
@@ -1,5 +1,9 @@
 import * as R from 'ramda'
-import styled, { ThemeProps as StyledThemeProps } from 'styled-components'
+import * as React from 'react'
+import styled, {
+  ThemeContext as StyledThemeContext,
+  ThemeProps as StyledThemeProps
+} from 'styled-components'
 
 import { ExpandableBoxTheme } from './components'
 
@@ -72,7 +76,14 @@ export type RendererThemeProps = StyledThemeProps<{
   rendererUi: RendererUiTheme
 }>
 
-export function createRendererUiElementTheme<T>(
+export function useRendererTheme(): {
+  renderer: RendererTheme
+  rendererUi: RendererUiTheme
+} {
+  return React.useContext(StyledThemeContext)
+}
+
+export function createRendererUiTheme<T>(
   createDefaultTheme: (theme: RendererTheme) => T
 ) {
   return (
@@ -85,6 +96,15 @@ export function createRendererUiElementTheme<T>(
     ) as unknown) as T
   }
 }
+export function useRendererUiTheme<T>(
+  key: keyof RendererUiTheme,
+  createDefaultTheme: RendererUiThemeFactory<T>
+) {
+  const theme = useRendererTheme()
+  return createRendererUiTheme(createDefaultTheme)(key, theme)
+}
+
+export type RendererUiThemeFactory<T> = (theme: RendererTheme) => T
 
 type DeepPartial<T> = {
   [P in keyof T]?: T[P] extends (infer U)[]